Adds pidfile support
This commit adds support for creating a pidfile on launch.
It adds switches "--pidfile" and "-Z"; the short option
is "-Z" because "p" and "P" were unavailable.

I was unable to determine a way to clean up the pidfile
on termination. So the pidfile created will end up
being stale and is up to the invoker to deal with.